fix: use code-splitting on routes
