+++ src/routes/index.lazy.tsx
@@ -2,7 +2,7 @@ import { Stack, Typography } from '@mui/material';
 
 import { Button, DEFAULT_LIGHT_PRIMARY_COLOR, GraaspLogo } from '@graasp/ui';
 
-import { Link, createFileRoute } from '@tanstack/react-router';
+import { Link, createLazyFileRoute } from '@tanstack/react-router';
 
 import { useAuth } from '@/AuthContext';
 import { LeftHeaderWrapper } from '@/components/header/LeftHeaderWrapper';
@@ -10,7 +10,7 @@ import { useAccountTranslation } from '@/config/i18n';
 import { ACCOUNT_HOME_PATH, LANDING_PAGE_PATH } from '@/config/paths';
 import { ACCOUNT } from '@/langs/constants';
 
-export const Route = createFileRoute('/')({
+export const Route = createLazyFileRoute('/')({
   component: Index,
 });
